requirements.txt
PyYAML==3.11
docopt==0.6.1
pykwalify==14.08
python-etcd==0.3.2
